services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
psycopg2-binary==2.9.9
pypdf==4.3.1
numpy==1.26.4
requests==2.32.3
